SalaryEvolution of user

+++ src/main/java/com/tlrh/gestion_tlrh_backend/dto/CollaborateurDto.java
@@ -13,7 +13,7 @@ public class CollaborateurDto {
     private String Bu;
     private String email;
 
-    private int SalaireActuel;
+    private Double SalaireActuel;
 
     private Integer managerRH;
