+++ src/main/java/com/tlrh/gestion_tlrh_backend/dto/UpdateBy3ActorsDto.java
@@ -7,7 +7,7 @@ public class UpdateBy3ActorsDto {
 
     private Integer matricule;
 
-    private int SalaireActuel;
+    private Double SalaireActuel;
 
     private Integer managerRH;
